+++ Paradigm/src/utopia/paradigm/color/HslLike.scala
@@ -192,29 +192,4 @@ trait HslLike[Repr <: HslLike[Repr]]
 	
 	def mapLuminosity(f: Double => Double) = withLuminosity(f(luminosity))
 	def mapDarkness(f: Double => Double) = withDarkness(f(darkness))
-	
-	/**
-	  * @param lightMod A lightening modifier > 0 where 2 is twice as bright, 1 keeps this color as is
-	  * @return A lightened version of this color
-	  */
-	@deprecated("Replaced with .lightenedBy and .lightened", "v1.2")
-	def lightened(lightMod: Double) = {
-		if (lightMod <= 0)
-			withLuminosity(0)
-		else {
-			val o = luminosity
-			withLuminosity(1 - (1 - o) / lightMod)
-		}
-	}
-	/**
-	  * @param darkMod A darkening modifier > 0 where 2 is twice as dark, 1 keeps this color as is
-	  * @return A darkened version of this color
-	  */
-	@deprecated("Replaced with .darkenedBy and .darkened", "v1.2")
-	def darkened(darkMod: Double) = {
-		if (darkMod <= 0)
-			withLuminosity(1)
-		else
-			timesLuminosity(1 / darkMod)
-	}
 }