+++ Reach/src/utopia/reach/component/label/text/ViewTextLabel.scala
@@ -10,6 +10,8 @@ import utopia.firmament.localization.{DisplayFunction, LocalizedString}
 import utopia.firmament.model.TextDrawContext
 import utopia.firmament.model.stack.StackInsets
 import utopia.flow.collection.immutable.Empty
+import utopia.flow.event.listener.ChangeListener
+import utopia.flow.event.model.ChangeResponse.Continue
 import utopia.flow.view.immutable.eventful.{Always, AlwaysFalse, AlwaysTrue, Fixed}
 import utopia.flow.view.template.eventful.{Changing, Flag}
 import utopia.genesis.graphics.Priority
@@ -264,16 +266,18 @@ class ViewTextLabel[+A](override val hierarchy: ComponentHierarchy, override val
 	// Revalidates and/or repaints this component on all text changes
 	textPointer.addListener { event =>
 		if (event.equalsBy { _.size })
-			repaint()
+			Continue.and { repaint() }
 		else
-			revalidate()
+			Continue.and { revalidate() }
 	}
 	// Style changes (color & alignment) also trigger a revalidation / repaint
 	stylePointer.addListenerWhile(hierarchy.linkedFlag) { event =>
 		if (event.equalsBy { _.color } || event.equalsBy { _.alignment })
-			repaint(Priority.Low)
+			Continue.and { repaint(Priority.Low) }
+		else
+			Continue
 	}
-	allowTextShrinkFlag.addListenerWhile(hierarchy.linkedFlag) { _ => revalidate() }
+	allowTextShrinkFlag.addListenerWhile(hierarchy.linkedFlag)(ChangeListener.triggerAfterEffect { revalidate() })
 	
 	
 	// IMPLEMENTED	-------------------------------------