+++ Reach/src/utopia/reach/container/wrapper/Swapper.scala
@@ -5,6 +5,7 @@ import utopia.firmament.drawing.immutable.CustomDrawableFactory
 import utopia.firmament.drawing.template.CustomDrawer
 import utopia.flow.collection.immutable.caching.cache.Cache
 import utopia.flow.collection.immutable.{Empty, Single}
+import utopia.flow.event.listener.ChangeListener
 import utopia.flow.util.Mutate
 import utopia.flow.view.immutable.eventful.Fixed
 import utopia.flow.view.mutable.eventful.SettableFlag
@@ -254,7 +255,7 @@ class Swapper[A](override val hierarchy: ComponentHierarchy, override val valueP
 	// INITIAL CODE	-------------------------------
 	
 	// Revalidates this container whenever content is swapped
-	contentPointer.addListenerWhile(hierarchy.linkedFlag) { _ => revalidate() }
+	contentPointer.addListenerWhile(hierarchy.linkedFlag)(ChangeListener.triggerAfterEffect { revalidate() })
 	
 	
 	// COMPUTED	-----------------------------------