+++ Paradigm/Changes.md
@@ -12,6 +12,7 @@
 - Modified `toString` implementations of **Vector2D**, **Vector3D**, **Point** and **Size**
 - Made some modifications to data type conversion priorities
 - **ColorSet** now extends **FromShadeFactory**
+- Removed **HslLike**'s `.darkened(Double)` and `.lightened(Double)` that were deprecated in v1.2
 
 ## v1.7.1 - 23.01.2025
 This update focuses on the **Distance** class, adding 2D vector representations with real life distance units.